chunker/requirements.txt
langchain
python-dotenv
openai
tiktoken
psycopg2-binary
pgvector